+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 1.0.0
 
 - Initial version.
+- Do not create records for blank rows in the CSV data.